+++ src/main/kotlin/Proto.kt
@@ -1,5 +1,5 @@
 data class Proto(
     val angle: Double,
-    val x: Int,
-    val y: Int,
+    val x: Double,
+    val y: Double,
 )
